feat: Add reward shaping class
