+++ gym_simplifiedtetris/envs/simplified_tetris_part_binary_env.py
@@ -1,3 +1,4 @@
+from typing import Optional, Sequence
 import numpy as np
 from gym import spaces
 
@@ -27,6 +28,11 @@ class SimplifiedTetrisPartBinaryEnv(SimplifiedTetrisStandardEnv):
             ),
             dtype=np.int,
         )
+
+    def __init__(
+        self, grid_dims: Sequence[int], piece_size: int, seed: Optional[int] = 8191
+    ):
+        super().__init__(grid_dims, piece_size, seed)
 
     def _get_obs(self) -> np.array:
         """
