+++ run_tests.py
@@ -7,10 +7,10 @@ from gym_simplifiedtetris.register import env_list
 
 def main() -> None:
     """
-    Checks if each env created conforms to the OpenAI Gym API.
-    The first observation is printed out for visual inspection.
-    Ten games are played using an agent that selects actions uniformly at random.
-    In every game, the reward received is validated and the env is rendered for visual inspection.
+    Checks if each env created conforms to the OpenAI Gym API. The first observation is printed
+    out for visual inspection. Ten games are played using an agent that selects actions uniformly
+    at random. In every game, the reward received is validated and the env is rendered for visual
+    inspection.
     """
     num_envs = len(env_list)
 
@@ -39,6 +39,7 @@ def main() -> None:
             ), f"Reward seen: {reward}"
 
             if num_episodes == 0 and is_first_move:
+                print(f"Reward range: {env.reward_range}")
                 print(f"First reward seen: {reward}")
                 print(f"Second observation given: {obs}")
                 is_first_move = False
